Api Municipio

+++ app/Http/Controllers/api/ComunaController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 use App\Models\Comuna;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator; 
+
 
 class ComunaController extends Controller
 {
@@ -16,9 +18,9 @@ class ComunaController extends Controller
     public function index()
     {
         $comunas = DB::table('tb_comuna')
-        ->join('tb_municipio','tb_comuna.muni_codi','=','tb_municipio.muni_codi')
-        ->select('tb_comuna.*', "tb_municipio.muni_nomb")
-        ->get();
+            ->join('tb_municipio', 'tb_comuna.muni_codi', '=', 'tb_municipio.muni_codi')
+            ->select('tb_comuna.*', "tb_municipio.muni_nomb")
+            ->get();
         return response()->json(['comunas' => $comunas], 200, [], JSON_PRETTY_PRINT);
     }
 
@@ -28,6 +30,17 @@ class ComunaController extends Controller
      */
     public function store(Request $request)
     {
+        $validate = Validator::make($request->all(), [
+            'comu_nomb' => ['required', 'max:30', 'unique'],
+            'muni_codi' => ['required', 'numeric', 'min:1']
+        ]);
+
+        if ($validate->fails()) {
+            return response()->json([
+                'msg' => 'Se produjo un error en la validación de la información',
+                'statusCode' => 400
+            ]);
+        }
         $comuna = new Comuna();
         $comuna->comu_nomb = $request->comu_nomb;
         $comuna->muni_codi = $request->muni_codi;
@@ -43,9 +56,12 @@ class ComunaController extends Controller
     public function show(string $id)
     {
         $comuna = Comuna::find($id);
+        if (is_null($comuna)) {
+            return abort(404);
+        }
         $municipios = DB::table('tb_municipio')
-        ->orderBy('muni_nomb')
-        ->get();
+            ->orderBy('muni_nomb')
+            ->get();
         return response()->json(['comunas' => $comuna, 'municipios' => $municipios], 200, [], JSON_PRETTY_PRINT);
     }
 
@@ -54,7 +70,21 @@ class ComunaController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $validate = Validator::make($request->all(), [
+            'comu_nomb' => ['required', 'max:30', 'unique'],
+            'muni_codi' => ['required', 'numeric', 'min:1']
+        ]);
+
+        if ($validate->fails()) {
+            return response()->json([
+                'msg' => 'Se produjo un error en la validación de la información',
+                'statusCode' => 400
+            ]);
+        }
         $comuna = Comuna::find($id);
+        if (is_null($comuna)) {
+            return abort(404);
+        }
         $comuna->comu_nomb = $request->comu_nomb;
         $comuna->muni_codi = $request->muni_codi;
         $comuna->save();
@@ -67,11 +97,14 @@ class ComunaController extends Controller
     public function destroy(string $id)
     {
         $comuna = Comuna::find($id);
+        if (is_null($comuna)) {
+            return abort(404);
+        }
         $comuna->delete();
         $comunas = DB::table('tb_comuna')
-        ->join('tb_municipio','tb_comuna.muni_codi','=','tb_municipio.muni_codi')
-        ->select('tb_comuna.*', "tb_municipio.muni_nomb")
-        ->get();
+            ->join('tb_municipio', 'tb_comuna.muni_codi', '=', 'tb_municipio.muni_codi')
+            ->select('tb_comuna.*', "tb_municipio.muni_nomb")
+            ->get();
         return response()->json(['comunas' => $comunas, 'success' => true], 200, [], JSON_PRETTY_PRINT);
     }
 }
